Move ObjEntry to separate module
