+++ src/main/scala/se/lu/nateko/cp/meta/services/sparql/index/FileSizeHierarchicalBitmap.scala
@@ -2,7 +2,7 @@ package se.lu.nateko.cp.meta.services.sparql.index
 
 import java.time.Instant
 import se.lu.nateko.cp.meta.core.algo.HierarchicalBitmap
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.ObjEntry
+import se.lu.nateko.cp.meta.services.sparql.magic.index.ObjEntry
 import scala.collection.IndexedSeq
 
 /**
