+++ src/main/scala/se/lu/nateko/cp/meta/services/sparql/index/SamplingHeightHierarchicalBitmap.scala
@@ -3,7 +3,7 @@ package se.lu.nateko.cp.meta.services.sparql.index
 import java.time.Instant
 import scala.annotation.tailrec
 import se.lu.nateko.cp.meta.core.algo.HierarchicalBitmap
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.ObjEntry
+import se.lu.nateko.cp.meta.services.sparql.magic.index.ObjEntry
 import scala.collection.IndexedSeq
 
 /**
