+++ src/main/scala/se/lu/nateko/cp/meta/services/sparql/magic/index/IndexData.scala
@@ -4,28 +4,26 @@ import scala.collection.mutable.ArrayBuffer
 import scala.collection.mutable.AnyRefMap
 import scala.collection.{IndexedSeq => IndSeq}
 import java.time.Instant
-
-import se.lu.nateko.cp.meta.utils.rdf4j.{===, toJava, Rdf4jStatement, asString}
-import se.lu.nateko.cp.meta.utils.{parseJsonStringArray, asOptInstanceOf, parseCommaSepList}
+import akka.event.LoggingAdapter
 import org.eclipse.rdf4j.model.IRI
 import org.eclipse.rdf4j.model.Literal
+import org.eclipse.rdf4j.model.Value
+import org.eclipse.rdf4j.model.Statement
 import org.eclipse.rdf4j.model.vocabulary.XSD
 import org.roaringbitmap.buffer.MutableRoaringBitmap
-
+import se.lu.nateko.cp.meta.utils.rdf4j.{===, toJava, Rdf4jStatement, asString}
+import se.lu.nateko.cp.meta.utils.{parseJsonStringArray, asOptInstanceOf, parseCommaSepList}
 import se.lu.nateko.cp.meta.services.sparql.index.*
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.ObjEntry
 import se.lu.nateko.cp.meta.core.crypto.Sha256Sum
 import se.lu.nateko.cp.meta.core.algo.HierarchicalBitmap
 import se.lu.nateko.cp.meta.core.algo.DatetimeHierarchicalBitmap
 import se.lu.nateko.cp.meta.core.algo.DatetimeHierarchicalBitmap.DateTimeGeo
+import se.lu.nateko.cp.meta.services.sparql.index.{FileSizeHierarchicalBitmap, SamplingHeightHierarchicalBitmap}
 import se.lu.nateko.cp.meta.services.sparql.index.StringHierarchicalBitmap.StringGeo
+import se.lu.nateko.cp.meta.services.linkeddata.UriSerializer.Hash
 import se.lu.nateko.cp.meta.services.CpVocab
 import se.lu.nateko.cp.meta.services.CpmetaVocab
-import se.lu.nateko.cp.meta.services.linkeddata.UriSerializer.Hash
-import org.eclipse.rdf4j.model.Value
-import akka.event.LoggingAdapter
 import se.lu.nateko.cp.meta.api.CloseableIterator
-import org.eclipse.rdf4j.model.Statement
 
 type StatementGetter = (subject: IRI | Null, predicate: IRI | Null, obj: Value | Null) => CloseableIterator[Statement]
 
@@ -49,11 +47,11 @@ class IndexData(nObjects: Int)(
 	val stats: AnyRefMap[StatKey, MutableRoaringBitmap] = AnyRefMap.empty,
 	val initOk: MutableRoaringBitmap = emptyBitmap
 ) extends Serializable {
-	def dataStartBm = DatetimeHierarchicalBitmap(DataStartGeo(objs))
-	def dataEndBm = DatetimeHierarchicalBitmap(DataEndGeo(objs))
-	def submStartBm = DatetimeHierarchicalBitmap(SubmStartGeo(objs))
-	def submEndBm = DatetimeHierarchicalBitmap(SubmEndGeo(objs))
-	def fileNameBm = StringHierarchicalBitmap(FileNameGeo(objs))
+	private def dataStartBm = DatetimeHierarchicalBitmap(DataStartGeo(objs))
+	private def dataEndBm = DatetimeHierarchicalBitmap(DataEndGeo(objs))
+	private def submStartBm = DatetimeHierarchicalBitmap(SubmStartGeo(objs))
+	private def submEndBm = DatetimeHierarchicalBitmap(SubmEndGeo(objs))
+	private def fileNameBm = StringHierarchicalBitmap(FileNameGeo(objs))
 
 	def boolBitmap(prop: BoolProperty): MutableRoaringBitmap = boolMap.getOrElseUpdate(prop, emptyBitmap)
